requirements.txt
lark_parser==0.7.8
pydot==1.4.1
networkx==2.4
